Fix warning
When adding 1 to the bit set we must cast 1 as a long value to about arithmetic overflow.

+++ InternalRepresentation/Main.cpp
@@ -17,7 +17,7 @@ int main() {
 
     // Add one bit to the bit set to give the two's compliment.
     Compliment = Invert;
-    Compliment = Compliment.to_ulong() + 1;
+    Compliment = Compliment.to_ulong() + (long)1;
 
     // IEEE conversion
     auto bitCast = std::bit_cast<uint32_t>(std::stof(arg));
